requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
charset-normalizer==3.4.2
colorama==0.4.6
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.3
propcache==0.3.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
starlette==0.46.2
Telethon==1.40.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
absl-py==2.2.2
altgraph==0.17.4
amqp==5.3.1
asgiref==3.8.1
astunparse==1.6.3
billiard==4.2.1
celery==5.5.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2
exceptiongroup==1.2.2
flatbuffers==25.2.10
gast==0.6.0
gevent==24.11.1
google-pasta==0.2.0
greenlet==3.2.0
grpcio==1.71.0
h5py==3.13.0
httpx==0.28.1
idna==3.10
keras==3.9.2
kombu==5.5.3
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mysqlclient==2.2.7
namex==0.0.8
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
pyarmor==9.1.3
pyarmor.cli.core==7.6.5
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.22.0
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pypiwin32==223
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==308
pywin32-ctypes==0.2.3
requests==2.32.3
rich==14.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
websocket_client==0.37.0
Werkzeug==3.1.3
wheel==0.45.1
WMI==1.5.1
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2